Player Verify

+++ proyecto/src/main/kotlin/com/manuel/proyecto/managers/PlayerManager.kt
@@ -1,5 +1,6 @@
 package com.manuel.proyecto.managers
 
+import com.manuel.proyecto.exception.BadRequest
 import com.manuel.proyecto.model.enity.Player
 import com.manuel.proyecto.persistence.PlayerDao
 import org.springframework.beans.factory.annotation.Autowired
@@ -18,16 +19,35 @@ class PlayerManager {
     }
 
     fun createPlayer(player: Player): Player {
-        //TODO verificar que el player tiene los datos correctos
+        player.idPlayer = null
+        if(player.name.isNullOrEmpty() || player.birthday == null || player.team == null || player.team?.idTeam == null){
+            throw BadRequest("Jugador invalido")
+        }
+
+        //TODO verificar id team
+        if(!dao.verifyIdTeam(player.team!!.idTeam!!)){
+            throw BadRequest("Equipo invalido")
+        }
+
         return dao.createPlayer(player)
     }
 
     fun updatePlayer(player: Player): Player {
-        //TODO verificar que el player tiene los datos correctos
+        if(player.idPlayer != null && player.name.isNullOrEmpty() || player.birthday == null || player.team == null || player.team?.idTeam == null){
+            throw BadRequest("Jugador invalido")
+        }
+
+        //TODO verificar id team
+        if(!dao.verifyIdTeam(player.team!!.idTeam!!)){
+            throw BadRequest("Equipo invalido")
+        }
         return dao.updatePlayer(player)
     }
 
     fun deletePlayer(player: Player): Boolean {
+        if(!dao.verifyIdPlayer(player.idPlayer!!)){
+            throw BadRequest("Jugador invalido")
+        }
         return dao.deletePlayer(player)
     }
